Update umbrel-app.yml
added "with best practices." towards the end of the sentence currently on Line three of the umbrel app yml for Lightning Node app, 

This add a nice ring to it because it denotes kind of a etiquette and is a nice term I like, 

Specifically, because it shows that Umbrel has done a lot of the Proof of Work necessary to build such a great product, and the team there I believe understand as well the, "sensible defaults," which is another great term I liked, 

Cheers friends, up to you, just a slight change.  Thank you.